webservice/requirements.txt
Flask
flask_sqlalchemy
flask_migrate
python-dotenv
psycopg2-binary
bcrypt
gunicorn
PyJWT
flask_hal